Extract Material models from config_repos to the top-level
