+++ server/webapp/WEB-INF/rails/webpack/models/config_repos/types.ts
@@ -19,18 +19,14 @@ import {Stream} from "mithril/stream";
 import * as stream from "mithril/stream";
 import {
   ConfigRepoJSON,
-  ConfigReposJSON, GitMaterialAttributesJSON, HgMaterialAttributesJSON, MaterialAttributesJSON,
-  MaterialJSON, MaterialModificationJSON,
-  P4MaterialAttributesJSON, ParseInfoJSON, SvnMaterialAttributesJSON, TfsMaterialAttributesJSON,
+  ConfigReposJSON, MaterialModificationJSON,
+  ParseInfoJSON,
 } from "models/config_repos/serialization";
+import {Material, Materials} from "models/materials/types";
 import {Errors} from "models/mixins/errors";
 import {applyMixins} from "models/mixins/mixins";
 import {ValidatableMixin} from "models/mixins/new_validatable_mixin";
-import {ErrorMessages} from "models/mixins/validatable";
 import {Configuration, PlainTextValue} from "models/shared/plugin_infos_new/plugin_settings/plugin_settings";
-import {EncryptedValue} from "views/components/forms/encrypted_value";
-
-const s = require("helpers/string-plus");
 
 //tslint:disable-next-line
 export interface ConfigRepo extends ValidatableMixin {
@@ -38,34 +34,6 @@ export interface ConfigRepo extends ValidatableMixin {
 
 //tslint:disable-next-line
 export interface LastParse extends ValidatableMixin {
-}
-
-//tslint:disable-next-line
-export interface Material extends ValidatableMixin {
-}
-
-//tslint:disable-next-line
-export interface MaterialAttributes extends ValidatableMixin {
-}
-
-//tslint:disable-next-line
-export interface GitMaterialAttributes extends ValidatableMixin {
-}
-
-//tslint:disable-next-line
-export interface SvnMaterialAttributes extends ValidatableMixin {
-}
-
-//tslint:disable-next-line
-export interface HgMaterialAttributes extends ValidatableMixin {
-}
-
-//tslint:disable-next-line
-export interface P4MaterialAttributes extends ValidatableMixin {
-}
-
-//tslint:disable-next-line
-export interface TfsMaterialAttributes extends ValidatableMixin {
 }
 
 export class ConfigRepo implements ValidatableMixin {
@@ -234,264 +202,6 @@ export class ParseInfo {
     }
   }
 }
-
-class Materials {
-  static fromJSON(material: MaterialJSON): Material {
-    return new Material(material.type, MaterialAttributes.deserialize(material));
-  }
-}
-
-export class Material implements ValidatableMixin {
-  type: Stream<string>;
-  attributes: Stream<MaterialAttributes>;
-
-  constructor(type?: string, attributes?: MaterialAttributes) {
-    this.attributes = stream(attributes);
-    this.type       = stream(type);
-    ValidatableMixin.call(this);
-    this.validateAssociated("attributes");
-  }
-
-  typeProxy(value?: any): string {
-    if (arguments.length > 0) {
-      const newType = value;
-      if (this.type() !== newType) {
-        this.attributes(MaterialAttributes.deserialize({
-                                                         type: newType,
-                                                         attributes: ({} as MaterialAttributesJSON)
-                                                       }));
-      }
-      this.type(newType);
-    }
-    return this.type();
-  }
-
-  materialUrl(): string {
-    // @ts-ignore
-    return this.type() === "p4" ? this.attributes().port() : this.attributes().url();
-  }
-}
-
-applyMixins(Material, ValidatableMixin);
-
-export abstract class MaterialAttributes implements ValidatableMixin {
-
-  autoUpdate: Stream<boolean>;
-
-  protected constructor(name?: string, autoUpdate?: boolean) {
-    this.autoUpdate = stream(autoUpdate);
-    ValidatableMixin.call(this);
-  }
-
-  static deserialize(material: MaterialJSON) {
-    switch (material.type) {
-      case "git":
-        return GitMaterialAttributes.fromJSON(material.attributes as GitMaterialAttributesJSON);
-      case "svn":
-        return SvnMaterialAttributes.fromJSON(material.attributes as SvnMaterialAttributesJSON);
-      case "hg":
-        return HgMaterialAttributes.fromJSON(material.attributes as HgMaterialAttributesJSON);
-      case "p4":
-        return P4MaterialAttributes.fromJSON(material.attributes as P4MaterialAttributesJSON);
-      case "tfs":
-        return TfsMaterialAttributes.fromJSON(material.attributes as TfsMaterialAttributesJSON);
-      default:
-        throw new Error(`Unknown material type ${material.type}`);
-    }
-  }
-
-  toJSON() {
-    const serialized                       = _.assign({}, this);
-    const password: Stream<EncryptedValue> = _.get(serialized, "password");
-
-    // remove the password field and setup the password serialization
-    if (password) {
-      // @ts-ignore
-      delete serialized.password;
-
-      if (password().isPlain() || password().isDirty()) {
-        return _.assign({}, serialized, {password: password().value()});
-      } else {
-        return _.assign({}, serialized, {encrypted_password: password().value()});
-      }
-    }
-
-    return serialized;
-  }
-}
-
-applyMixins(MaterialAttributes, ValidatableMixin);
-
-export class GitMaterialAttributes extends MaterialAttributes {
-  url: Stream<string>;
-  branch: Stream<string>;
-
-  constructor(name?: string, autoUpdate?: boolean, url?: string, branch?: string) {
-    super(name, autoUpdate);
-    this.url    = stream(url);
-    this.branch = stream(branch);
-    this.validatePresenceOf("url");
-  }
-
-  static fromJSON(json: GitMaterialAttributesJSON) {
-    const gitMaterialAttributes = new GitMaterialAttributes(json.name, json.auto_update, json.url, json.branch);
-    gitMaterialAttributes.errors(new Errors(json.errors));
-    return gitMaterialAttributes;
-  }
-}
-
-applyMixins(GitMaterialAttributes, ValidatableMixin);
-
-interface PasswordLike {
-  cipherText?: string;
-  plainText?: string;
-}
-
-function plainOrCipherValue(passwordLike: PasswordLike) {
-  if (passwordLike.cipherText) {
-    return new EncryptedValue({cipherText: s.defaultToIfBlank(passwordLike.cipherText, "")});
-  } else {
-    return new EncryptedValue({clearText: s.defaultToIfBlank(passwordLike.plainText, "")});
-  }
-}
-
-export class SvnMaterialAttributes extends MaterialAttributes implements ValidatableMixin {
-  url: Stream<string>;
-  checkExternals: Stream<boolean>;
-  username: Stream<string>;
-  password: Stream<EncryptedValue>;
-
-  constructor(name?: string,
-              autoUpdate?: boolean,
-              url?: string,
-              checkExternals?: boolean,
-              username?: string,
-              password?: string,
-              encryptedPassword?: string) {
-    super(name, autoUpdate);
-    this.url            = stream(url);
-    this.checkExternals = stream(checkExternals);
-    this.username       = stream(username);
-    this.password       = stream(plainOrCipherValue({plainText: password, cipherText: encryptedPassword}));
-    this.validatePresenceOf("url");
-  }
-
-  static fromJSON(json: SvnMaterialAttributesJSON) {
-    const svnMaterialAttributes = new SvnMaterialAttributes(json.name,
-                                                            json.auto_update,
-                                                            json.url,
-                                                            json.check_externals,
-                                                            json.username,
-                                                            json.password,
-                                                            json.encrypted_password);
-    svnMaterialAttributes.errors(new Errors(json.errors));
-    return svnMaterialAttributes;
-  }
-}
-
-applyMixins(SvnMaterialAttributes, ValidatableMixin);
-
-export class HgMaterialAttributes extends MaterialAttributes {
-  url: Stream<string>;
-
-  constructor(name?: string, autoUpdate?: boolean, url?: string) {
-    super(name, autoUpdate);
-    this.url = stream(url);
-    this.validatePresenceOf("url");
-  }
-
-  static fromJSON(json: HgMaterialAttributesJSON) {
-    const hgMaterialAttributes = new HgMaterialAttributes(json.name, json.auto_update, json.url);
-    hgMaterialAttributes.errors(new Errors(json.errors));
-    return hgMaterialAttributes;
-  }
-}
-
-applyMixins(HgMaterialAttributes, ValidatableMixin);
-
-export class P4MaterialAttributes extends MaterialAttributes {
-  port: Stream<string>;
-  useTickets: Stream<boolean>;
-  view: Stream<string>;
-  username: Stream<string>;
-  password: Stream<EncryptedValue>;
-
-  constructor(name?: string,
-              autoUpdate?: boolean,
-              port?: string,
-              useTickets?: boolean,
-              view?: string,
-              username?: string,
-              password?: string,
-              encryptedPassword?: string) {
-    super(name, autoUpdate);
-    this.port       = stream(port);
-    this.useTickets = stream(useTickets);
-    this.view       = stream(view);
-    this.username   = stream(username);
-    this.password   = stream(plainOrCipherValue({plainText: password, cipherText: encryptedPassword}));
-    this.validatePresenceOf("view");
-    this.validatePresenceOf("port", {message: ErrorMessages.mustBePresent("Host and Port")});
-  }
-
-  static fromJSON(json: P4MaterialAttributesJSON) {
-    const p4MaterialAttributes = new P4MaterialAttributes(json.name,
-                                                          json.auto_update,
-                                                          json.port,
-                                                          json.use_tickets,
-                                                          json.view,
-                                                          json.username,
-                                                          json.password,
-                                                          json.encrypted_password);
-    p4MaterialAttributes.errors(new Errors(json.errors));
-    return p4MaterialAttributes;
-  }
-}
-
-applyMixins(P4MaterialAttributes, ValidatableMixin);
-
-export class TfsMaterialAttributes extends MaterialAttributes {
-  url: Stream<string>;
-  domain: Stream<string>;
-  projectPath: Stream<string>;
-  username: Stream<string>;
-  password: Stream<EncryptedValue>;
-
-  constructor(name?: string,
-              autoUpdate?: boolean,
-              url?: string,
-              domain?: string,
-              projectPath?: string,
-              username?: string,
-              password?: string,
-              encryptedPassword?: string) {
-    super(name, autoUpdate);
-    this.url         = stream(url);
-    this.domain      = stream(domain);
-    this.projectPath = stream(projectPath);
-    this.username    = stream(username);
-    this.password    = stream(plainOrCipherValue({plainText: password, cipherText: encryptedPassword}));
-    this.validatePresenceOf("url");
-    this.validatePresenceOf("projectPath");
-    this.validatePresenceOf("username");
-    this.validatePresenceOfPassword("password");
-  }
-
-  static fromJSON(json: TfsMaterialAttributesJSON) {
-    const tfsMaterialAttributes = new TfsMaterialAttributes(json.name,
-                                                            json.auto_update,
-                                                            json.url,
-                                                            json.domain,
-                                                            json.project_path,
-                                                            json.username,
-                                                            json.password,
-                                                            json.encrypted_password);
-    tfsMaterialAttributes.errors(new Errors(json.errors));
-    return tfsMaterialAttributes;
-  }
-}
-
-applyMixins(TfsMaterialAttributes, ValidatableMixin);
 
 const HUMAN_NAMES_FOR_MATERIAL_ATTRIBUTES: { [index: string]: string } = {
   autoUpdate: "Auto update",
